Adding a CI
The essence of CI is providing rapid feedback to developers about the quality of the code they write
so they can find and resolve issues quickly.

We're going to be adding flake8.yml to our repository, so we'd like to at least be able to
verify that the code we add conforms to proper Python syntax and coding standards. To do this we
will use the "Flake8" tool and set things up so that it scans our code automatically as we send PRs
and provides feedback on the PR review screen.